+++ R/CA.MI.R
@@ -24,7 +24,7 @@
 #' @references
 #' Iaconangelo, C.(2017). \emph{Uses of classification error probabilities in the three-step approach to estimating cognitive diagnosis models}. (Unpublished doctoral dissertation). New Brunswick, NJ: Rutgers University.
 #'
-#' Kreitchmann, R. S., de la Torre, J., Sorrel, M. A., Nájera, P., & Abad, F. J. (2022). Improving reliability estimation in cognitive diagnosis modeling. [Manuscript submitted for publication].
+#' Kreitchmann, R. S., de la Torre, J., Sorrel, M. A., Nájera, P., & Abad, F. J. (2022). Improving reliability estimation in cognitive diagnosis modeling. \emph{Behavior Research Methods}. https://doi.org/10.3758/s13428-022-01967-5
 #'
 #' Ma, W., & de la Torre, J. (2020). GDINA: An R package for cognitive diagnosis modeling. \emph{Journal of Statistical Software}, \emph{93}(14). https://doi.org/10.18637/jss.v093.i14
 #'
@@ -40,47 +40,75 @@
 #' dat <- sim10GDINA$simdat[1:100,]
 #' Q <- sim10GDINA$simQ
 #' fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
-#' ca.mi <- CA.MI(fit, R = 100) # R > 500 is recommended
+#' ca.mi <- CA.MI(fit)
 #' ca.mi
 #' }
-CA.MI <- function(fit, what = "EAP", R = 500, n.cores = 1, verbose = TRUE, seed = NULL){
 
+CA.MI <- function(fit, what = "EAP", R = 500, n.cores = 1, verbose = TRUE, seed = NULL){
+  
   if(!inherits(fit, "GDINA")){stop("Error in CA.MI: fit must be of class 'GDINA'.")}
   if(!what %in% c("EAP", "MAP", "MLE")){stop("Error in CA.MI: what must be either 'EAP', 'MAP', or 'MLE'.")}
-  if(any(apply(fit$options$dat, 2, stats::sd, na.rm = TRUE) == 0)){stop("Error in CA.MI: The data must not contain constant responses for an item.")}
+  if(any(apply(fit$options$dat, 2, stats::sd, na.rm = T) == 0)){stop("Error in CA.MI: The data must not contain constant responses for an item.")}
   if((!is.numeric(R) & !is.double(R)) | length(R) > 1){stop("Error in CA.MI: R must be a unique numeric value.")}
   if(R < 1){stop("Error in CA.MI: R must be greater than 0.")}
   if((!is.numeric(n.cores) & !is.double(n.cores)) | length(n.cores) > 1){stop("Error in CA.MI: n.cores must be a unique numeric value.")}
   if(n.cores < 1){stop("Error in CA.MI: n.cores must be greater than 0.")}
   if(!is.logical(verbose)){stop("Error in CA.MI: verbose must be logical.")}
   if(!is.null(seed)){if((!is.numeric(seed) & !is.double(seed)) | length(seed) > 1){stop("Error in CA.MI: seed must be a unique numeric value.")}}
-
+  if(extract(fit, "ngroup") != 1) {stop("Error in CA.MI: only applicable for single group analysis.")}
+  
   if(is.null(seed)){seed <- sample(1:100000, 1)}
   set.seed(seed)
-
+  
   GDINA.options <- formals(GDINA::GDINA)
   GDINA.options <- GDINA.options[-c(1, 2, length(GDINA.options))]
   tmp <- as.list(fit$extra$call)[-c(1:3)]
   GDINA.options[names(GDINA.options) %in% names(tmp)] <- tmp
   GDINA.options$verbose <- 0
   GDINA.options$att.dist <- "fixed"
-  GDINA.options$att.prior <- fit$struc.parm
+  GDINA.options$model <- fit$model
   GDINA.options$control <- list(maxitr = rep(0, nrow(fit$options$Q)))
-
+  
   boot.sampling.dist <- bootSE.parallel(fit, bootsample = R, n.cores = n.cores, verbose = verbose, seed = seed)
-
+  
   posterior.R <- array(NA, dim = c(dim(t(fit$technicals$logposterior.i)), R))
   if(verbose){cat("\n", "\n")}
   for(r in 1:R){
     if(verbose & r %% 100 == 0){cat("Imputing Item Parameters: Iteration", r, "of", R, "\r")}
     GDINA.options$catprob.parm <- boot.sampling.dist$boot.est$itemprob[[r]]
+    GDINA.options$att.prior <- boot.sampling.dist$boot.est$lambda[[r]]
     fit.tmp <- do.call(GDINA::GDINA, c(list(dat = fit$options$dat, fit$options$Q), GDINA.options))
     posterior.R[,,r] <- as.matrix(exp(t(fit.tmp$technicals$logposterior.i)))
   }
-
+  
   posterior <- apply(posterior.R, 1:2, mean)
   fit.MI <- fit
   fit.MI$technicals$logposterior.i <- t(log(posterior))
-  res <- GDINA::CA(fit.MI, what = what)
+  
+  p_c <- GDINA::extract(fit.MI, "posterior.prob")
+  pp <- GDINA::personparm(fit, what = what)
+  if (what == "MAP" || what == "MLE") {
+    if (any(pp[, ncol(pp)])) 
+      warning(paste0(what, " estimates for some individuals have multiple modes.", 
+                     collapse = ""), call. = FALSE)
+    pp <- as.matrix(pp[, -ncol(pp)])
+  }
+  mp <- GDINA::personparm(fit.MI, what = "mp")
+  patt <- GDINA::extract(fit.MI, "attributepattern")
+  gr <- GDINA:::matchMatrix(patt, pp)
+  pseudo.gr <- setdiff(seq(nrow(patt)), unique(gr))
+  gr <- c(gr, pseudo.gr)
+  lab <- apply(patt, 1, paste0, collapse = "")
+  post <- cbind(exp(t(GDINA:::indlogPost(fit.MI))), matrix(0, nrow(patt), 
+                                                           length(pseudo.gr)))
+  CCM <- GDINA:::aggregateCol(post, gr)/c(GDINA::extract(fit.MI, "nobs") * 
+                                            p_c)
+  tau_c <- diag(CCM)
+  tau <- sum(tau_c * c(p_c))
+  tau_k <- colMeans(pp * mp + (1 - pp) * (1 - mp))
+  names(tau_c) <- rownames(CCM) <- colnames(CCM) <- lab
+  res <- list(tau = tau, tau_l = tau_c, tau_k = tau_k, CCM = CCM)
+  class(res) <- "CA"
+  
   return(res)
 }